Minor Updates to ffmpeg_reader
Support proper sub-second seeking. Needs optimization

+++ modules/filmstro_ffmpeg/filmstro_ffmpeg_FFmpegVideoReader.h
@@ -289,7 +289,7 @@ private:
 
     bool        looping;
 
-    int         sampleRate;
+    double      sampleRate;
 
     double      framesPerSec;
